fix(ci): python-full-coverage — split oversized setup CLI test file under line limit
test_setup_commands.py grew to 1750 lines, tripping the maintainability
guardrail (test_first_party_code_files_stay_under_line_limit, 1500-line cap).
Extract shared harness fixtures/helpers into _setup_commands_shared.py (re-exported
via conftest.py) and move the T08 client-integration dispatch tests into
test_setup_commands_client.py. No tests changed: 66 setup tests preserved (45 + 21).
